Fix an issue with the Fetch Register Clock
First the clock signal was not being routed to the Fetch Register.  This
was quickly resolved.

Then, the Instruction Suppress signal really needs to be inverted before
it is ANDed with the clock.  This was sinply forgotten as I had it in my
notes to make this happen.  This was also quickly corrected.

The clock is now working, latching the signals on Address Bus 1, which
happens 1 cycle after the PC latches the same signals.  This is
appropriate timing (at least for the moment).