Fix the name of the bus with multiple asserts
The code was properly detecting that multiple things were asserting onto
a bus, but had no knowledge of that bus's name.

This fix addresses this by requiring a name of a bus to be specified
when creating the bus, and then passing that name to `SetObjectName()`.

Of course, this was not a defaulted parameter for Bus creation, so every
instance had to be modified to address this concern.