Correct an issue with the Control ROM Control Mod
The Control ROM Control Module MUXes in the emulator were all tied to
the least significant nibble on the instruction coming from the
instruction register.

This is obviously not correct.

This commit updates the Control ROM Control Module to properly
distribute those signals so that each MUX input pin has its own unique
and correct signal.

I reviewed the schematic and there are no issues there, so I did that
translation for the schematic without validating against the emulator
code.