Fix the Control ROM always having signals on
This was caused by the SRAM (AS6C62256 IC) not signaling the updates
properly for all changes in output.  To overcome this, the code was
changed to update all outputs and signal that every line changed.

This also had the effect of cleaning up the concern that a NOP was not
asserted when the Instruction Suppress signal was asserted.

Some additional debug code was found and removed for clarity.