Add a test ROM build for the Control ROM Modules
This commit adds a control ROM for testing the Control ROM Modules.
These will be manual tests, but the contents of the ROM are laid out
such that output values are easily produced and repeatable, with the
first and last bytes set to something that is not typical for those
locations -- making it easy to confirm that the ROM is being read into
SRAM properly.